Update maker.py
trailing newlines on the last item causes some issues in the program.  for some reason when having an empty item in the list when running ppmaster.py, the correct answer is not shown for that item.  for example in {'key':['value','']}, the program would not output the correct answer which is ['value',''] and instead just skip printing stuff and waiting for user input.  having an rstrip doesnt really fix the problem but just remove it